src/backend/requirements.txt
fastapi==0.95.2
python-jose
uvicorn==0.22.0
fastapi_jwt_auth
decouple
requests==2.30.0
PyJWT==2.7.0
httpx==0.24.1
protobuf==4.23.1
idna
certifi==2023.5.7
urllib3==2.0.2
minio==7.1.15
yoyo-migrations~=8.2.0
psycopg2
SQLAlchemy==2.0.15
python-multipart
filetype~=1.2.0
pydantic~=1.10.7
starlette==0.27.0
email-validator==2.0.0.post2
flatten_json
cryptography
base32hex
jwt
jsonpath-ng
python-keycloak
